requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-cruds-adminlte==0.0.13
django-select2==6.3.1
djangoajax==3.0.3
djangorestframework==3.9.0
easy-thumbnails==2.5
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.20.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
